Refactor of view layer
