+++ CodingTracker.kkvzx/Views/helpers/Range.cs
